Added a test case for Persistent Reserve Out.
If a TransportID with an unknown initiator name is specified in the REGISTER service action, the TransportID will no longer be registered.
In anticipation of using the target node ID as a relative target port identifier in the future, the representation type of the target node ID has been changed from uint32 to uint16.

+++ products/Client/CommandRunner.fs
@@ -1488,9 +1488,9 @@ type CommandRunner( m_Messages : StringTable, m_InFile : TextReader, m_OutFile :
             | :? ConfNode_Target as x ->
                 match entNameUp with
                 | "ID" ->
-                    let r, v = UInt32.TryParse entValue
+                    let r, v = UInt16.TryParse entValue
                     if not r then
-                        m_Messages.GetMessage( "CMDMSG_PARAMVAL_DATATYPE_MISMATCH", "uint32" )
+                        m_Messages.GetMessage( "CMDMSG_PARAMVAL_DATATYPE_MISMATCH", "uint16" )
                         |> this.Output 0
                         return Some ( ss, cc, cn )
                     else
